Change error message - Connection failed
This commit changes the error message shown to the user when the client fails to connect to websockify. 
In particular, the message suggests the user to reload the page, since the error is most often due to the expiration of the authentication cookie and reloading the page solves it.
Unfortunately, it seems not possible to catch the 302 Redirect status code from javascript (due to security restrictions associated with websockets). Hence an automatized solution seems difficult to be implemented.